added tensor serialization (#68)
* added tensor serialization

* removed lateinit initialization, SyftModel integeration into job

* reformat code

* cleared duplicate http protocol

* addressed comments and rebasing

* remove message processor test

* removed pytorch tensor wrapper since we wont perform computation in kotlin

* added syftmodel test

* updated order of schedulers

+++ syftlib/src/main/java/org/openmined/syft/networking/datamodels/ClientConfig.kt
@@ -11,7 +11,7 @@ data class ClientConfig(
     @SerialName("version")
     val modelVersion: String,
     @SerialName("batch_size")
-    val batchSize: Int,
+    val batchSize: Long,
     val lr: Float,
     @SerialName("max_updates")
     val maxUpdates: Int
